Added v13
v13 – Merged contact and results pages into one. Contact.php displays the contact form until the "Send" button is clicked. Then, contact.php sends the data to itself (method: POST) and the page changes to display the data. The functional advantage of this is that the form data—the account credentials and personal information—will no longer be exposed in the URL.